Roulette途中

+++ Assets/Script/DataReader/CharacterData.cs
@@ -14,7 +14,7 @@ public class CharacterData : MonoBehaviour
         id = this.gameObject.GetComponentInParent<Character_Present_Data>()._ID;
         for (int i = 0; i < data.Character.Count; i++)
         {
-            if (data.Character[i].ID == id)
+            if ((int)data.Character[i].ID == id)
             {
                 CharacterStatus = data.Character[i];
                 break;
